refactors CustomButton is now a component.
